refactor - Merged GameMessage functions to OutputView
