+++ src/main/java/racingcar/views/InputView.java
@@ -13,7 +13,7 @@ public class InputView {
     public static String[] inputCarList(Scanner sc) {
         String[] carList;
 
-        GameMessages.printAskInputCarListMessage();
+        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
 
         try {
             carList = inputStringArray(sc);
@@ -27,7 +27,7 @@ public class InputView {
     public static Integer inputTurn(Scanner sc) {
         Integer turn;
 
-        GameMessages.printAskTurnNumberInputMessage();
+        System.out.println("시도할 회수는 몇회인가요?");
 
         try {
             turn = inputInteger(sc);
